refactor(core): rollup imported exports from core module
