+++ app/scripts/modules/core/account/index.ts
@@ -1,0 +1,2 @@
+export * from './account.service';
+export * from './AccountTag';
