+++ app/scripts/modules/core/api/index.ts
@@ -1,0 +1,1 @@
+export * from './api.service';
