+++ app/scripts/modules/core/cancelModal/index.ts
@@ -1,0 +1,1 @@
+export * from './cancelModal.service';
