+++ app/scripts/modules/core/ci/index.ts
@@ -1,0 +1,1 @@
+export * from './igor.service';
