+++ app/scripts/modules/core/cloudProvider/index.ts
@@ -1,0 +1,3 @@
+export * from './providerSelection/providerSelection.service';
+export * from './cloudProvider.registry';
+export * from './CloudProviderLogo';
