+++ app/scripts/modules/core/cluster/index.ts
@@ -1,0 +1,2 @@
+export * from './cluster.service';
+export * from './task.matcher';
