+++ app/scripts/modules/core/confirmationModal/index.ts
@@ -1,0 +1,1 @@
+export * from './confirmationModal.service';
