+++ app/scripts/modules/core/delivery/index.ts
@@ -1,0 +1,2 @@
+export * from './service/execution.service';
+export * from './executionGroup/execution/Execution';
