+++ app/scripts/modules/core/entityTag/index.ts
@@ -1,0 +1,3 @@
+export * from './entityTags.read.service';
+export * from './entityTags.write.service';
+export * from './clusterTargetBuilder.service';
