+++ app/scripts/modules/core/filterModel/index.ts
@@ -1,0 +1,1 @@
+export * from './FilterTags';
