+++ app/scripts/modules/core/help/index.ts
@@ -1,0 +1,1 @@
+export * from './helpContents.registry';
