+++ app/scripts/modules/core/history/index.ts
@@ -1,0 +1,1 @@
+export * from './recentHistory.service';
