+++ app/scripts/modules/core/image/index.ts
@@ -1,0 +1,1 @@
+export * from './image.reader';
