+++ app/scripts/modules/core/instance/index.ts
@@ -1,0 +1,2 @@
+export * from './instance.read.service';
+export * from './instance.write.service';
