+++ app/scripts/modules/core/loadBalancer/index.ts
@@ -1,0 +1,2 @@
+export * from './loadBalancer.read.service';
+export * from './loadBalancer.write.service';
