+++ app/scripts/modules/core/modal/index.ts
@@ -1,0 +1,3 @@
+export * from './buttons/SubmitButton';
+export { V2_MODAL_WIZARD_COMPONENT, V2ModalWizard } from './wizard/v2modalWizard.component';
+export { WizardPage, V2_MODAL_WIZARD_SERVICE, V2ModalWizardService, WizardPageState } from './wizard/v2modalWizard.service';
