+++ app/scripts/modules/core/naming/index.ts
@@ -1,0 +1,1 @@
+export * from './naming.service';
