+++ app/scripts/modules/core/network/index.ts
@@ -1,0 +1,1 @@
+export * from './network.read.service';
