+++ app/scripts/modules/core/orchestratedItem/index.ts
@@ -1,0 +1,1 @@
+export * from './orchestratedItem.transformer';
