+++ app/scripts/modules/core/overrideRegistry/index.ts
@@ -1,0 +1,1 @@
+export * from './override.registry';
