+++ app/scripts/modules/core/pageTitle/index.ts
@@ -1,0 +1,1 @@
+export * from './pageTitle.service';
