+++ app/scripts/modules/core/pipeline/config/stages/templates/index.ts
@@ -1,0 +1,5 @@
+export const destroyAsgExecutionDetailsTemplate = require('core/pipeline/config/stages/destroyAsg/templates/destroyAsgExecutionDetails.template.html');
+export const disableAsgExecutionDetailsTemplate = require('core/pipeline/config/stages/disableAsg/templates/disableAsgExecutionDetails.template.html');
+export const enableAsgExecutionDetailsTemplate = require('core/pipeline/config/stages/enableAsg/templates/enableAsgExecutionDetails.template.html');
+export const shrinkClusterExecutionDetailsTemplate = require('core/pipeline/config/stages/shrinkCluster/templates/shrinkClusterExecutionDetails.template.html');
+export const addExtendedAttributesTemplate = require('core/pipeline/config/stages/bake/modal/addExtendedAttribute.html');
