+++ app/scripts/modules/core/presentation/index.ts
@@ -1,0 +1,3 @@
+export * from './HoverablePopover';
+export * from './Tooltip';
+export * from './LabelComponent';
