+++ app/scripts/modules/core/scheduler/index.ts
@@ -1,0 +1,1 @@
+export * from './scheduler.factory';
