+++ app/scripts/modules/core/search/index.ts
@@ -1,0 +1,3 @@
+export * from './search.service';
+export * from './infrastructure/infrastructureSearch.service';
+export * from './searchResult/searchResultFormatter.registry';
