+++ app/scripts/modules/core/securityGroup/index.ts
@@ -1,0 +1,2 @@
+export * from './securityGroupReader.service';
+export * from './securityGroupWriter.service';
