+++ app/scripts/modules/core/serverGroup/index.ts
@@ -1,0 +1,8 @@
+export * from './serverGroupReader.service';
+export * from './serverGroupWriter.service';
+export * from './details/serverGroupWarningMessage.service';
+export * from './metrics/cloudMetrics.read.service';
+export * from './configure/common/serverGroupCommandBuilder.service';
+export * from './configure/common/serverGroupCommandRegistry.provider';
+
+export const userDataTemplate = require('./details/userData.html');
