+++ app/scripts/modules/core/serviceAccount/index.ts
@@ -1,0 +1,1 @@
+export * from './serviceAccount.service';
