+++ app/scripts/modules/core/subnet/index.ts
@@ -1,0 +1,1 @@
+export * from './subnet.read.service';
